Mobile Track 0: share JSON-safety (parseToolArgs + external-ref); fix silent empty-note write
Moves the two pure JSON-safety modules from the desktop PR into @cairn/shared so
desktop and mobile share one implementation, and adopts them on mobile.

- shared/chat/parse-tool-args.ts + external-ref.ts (moved from electron/lib;
  electron keeps thin re-export shims so desktop imports are unchanged). Tests
  moved alongside; run in the root vitest 'shared/**' project.
- mobile providers/apple.ts: the on-device tool path did
  JSON.parse(inputJson) catch {} = {} — a malformed ensure_note/patch_note
  (markdown content with raw newlines) SILENTLY ran with {} and wrote an EMPTY
  note. Now uses shared parseToolArgs (strict-first + lossless repair) and, on
  genuine parse failure, returns an error the model can re-issue against instead
  of destroying content. Display parse in the stream uses the same tolerant path.
- mobile chat: web/MCP/service tool results now yield a tappable external-link
  chip. ToolCall gains externalRef; chat/index.tsx extracts it via shared
  extractExternalRef (only when there's no in-app note/card ref); ToolTrail
  renders an https-guarded chip opening via Linking.openURL.

Verify: root 1373 pass/12 skip; type-check:all + lint green; mobile type-check
+ lint green.
