+++ mobile/app/(tabs)/chat/index.tsx
@@ -24,6 +24,7 @@ import { resetAppleSession } from "@/chat/providers/apple";
 import type { UIMessage, ChatUsage } from "@/chat/providers/types";
 import { ContextRing } from "@/components/ContextRing";
 import { toolRef } from "@cairn/shared/chat/tool-ref";
+import { extractExternalRef } from "@cairn/shared/chat/external-ref";
 import { loadInitialChat, type UiMessage } from "@/chat/history";
 import { useChatScroll } from "@/chat/useChatScroll";
 import { MessageBubble } from "@/components/chat/MessageBubble";
@@ -139,7 +140,13 @@ export default function ChatScreen() {
           patchAssistant({ reasoning: reasoningAcc });
         } else if (e.type === "tool" && e.tool) {
           const ok = !(e.result && typeof e.result === "object" && "error" in (e.result as object));
-          toolTrail.push({ tool: e.tool, ok, ref: toolRef(e.tool, e.args, e.result) });
+          // A note/card ref opens in-app; otherwise try for a linkable external
+          // URL (web-search hit, docs page, …) the chip can open in the browser.
+          const ref = toolRef(e.tool, e.args, e.result);
+          const externalRef = ref
+            ? undefined
+            : extractExternalRef(typeof e.result === "string" ? e.result : JSON.stringify(e.result ?? null));
+          toolTrail.push({ tool: e.tool, ok, ref, externalRef });
           patchAssistant({ tools: [...toolTrail] });
           haptics.impact(); // agent ran a tool
         } else if (e.type === "final" && e.usage) {
