+++ shared/chat/parse-tool-args.test.ts
